Store build config information in lib/bup/config/*
Instead of autogenerating _version.py, create and maintain individual
data files in lib/bup/config, starting with mandir and version.  The
former indicates the absolute path to the installed manpages, and the
latter contains the information previously in _version.py.  Add a new
config module to provide access to the data.

Point mandir to Documentation in the build tree, and to
MANPATH in each install tree, unless the manpages weren't generated.  In
that case remove mandir from the install tree, so that "bup help ..."
can detect and describe the problem.

Adjust "bup help" to prepend mandir to the MANPATH before invoking man
so that it always presents the manpages that correspond to the running
executable.

Adjust "bup version" to use config.version, and simplify the code.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>